% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mto_predictors.R
\name{ca_predictor}
\alias{ca_predictor}
\title{Car predictor}
\usage{
ca_predictor(data, model, fix = NULL, cc = NULL, return_vars = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with data to predict probability of owning a car}

\item{model}{Apollo model}

\item{fix}{Named vector. Fix coefficients to certain values}

\item{cc}{Calibration constant (cutoff parameter)}
}
\value{
\code{data.frame} with probability vectors
}
\description{
Car predictor
}
\seealso{
\code{\link[=base_predictor]{base_predictor()}}
}
