% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_predictor.R
\name{ga_predictor}
\alias{ga_predictor}
\title{GA predictor}
\usage{
ga_predictor(data, model, fix = NULL, cc = NULL, return_vars = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with data to predict probability of having GA subscription}

\item{model}{Apollo model}

\item{fix}{Named vector. Fix coefficients to certain values}

\item{cc}{Calibration constant (cutoff parameter\link{s})}
}
\value{
\code{data.frame} with probability vectors
}
\description{
GA predictor
}
