% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mto_predictors.R
\name{ht_predictor}
\alias{ht_predictor}
\title{Half-fare card (HT) predictor}
\usage{
ht_predictor(data, model, fix = NULL, cc = NULL, return_vars = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with data to predict probability of owning a HT}

\item{model}{Apollo model}

\item{fix}{Named vector. Fix coefficients to certain values}

\item{cc}{Calibration constant (cutoff parameter)}

\item{return_vars}{Boolean. If \code{TRUE} returns vars as required by \code{data}}
}
\value{
\code{data.frame} with probability vectors
}
\description{
Half-fare card (HT) predictor
}
\seealso{
\code{\link[=base_predictor]{base_predictor()}}
}
