% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfh_predictors.R
\name{wfh_predictor}
\alias{wfh_predictor}
\title{Home office predictor}
\usage{
wfh_predictor(data, apollo_list, fix = NULL, cc = NULL, return_vars = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with data to predict probability for selection and frequency model components}

\item{apollo_list}{Here we also need to have access to the \code{apollo_probabilities()} function}

\item{fix}{Passed to \code{apollo_fixed}}

\item{cc}{Named vector. One or multiple of \code{c}, \code{tau1}, \code{tau2}, \code{tau3}, \code{tau4}, \code{tau5}}

\item{return_vars}{Boolean. If \code{TRUE} returns vars as required by \code{data}}
}
\value{
List of \code{data.frame}s. One for selection component and one for frequency component
}
\description{
Home office predictor
}
